FIX Adds random_seed as a plugin
