+++ setup.cfg
@@ -10,6 +10,7 @@ addopts =
     --doctest-modules
     --disable-pytest-warnings
     --color=yes
+    -p sklearn.tests.random_seed
     -rN
 
 filterwarnings =
